Version 0.5.0 Parallelled lcpm More error messages and documentation Adding memory efficient options in several functions as the default Updated examples accordingly

+++ src/normalisr/binnet.py
@@ -123,7 +123,7 @@ def bh(pv,weight=None):
 		pv2[xi]=min(pv2[xi],pv2[xi+1])
 
 	#Recover index
-	ans=pv2[ids].astype(pv.dtype)
+	ans=pv2[ids].astype(pv.dtype,copy=False)
 	assert ans.shape==pv.shape
 	return ans
